+++ src/clj/lambdacd/stepresults/flatten.clj
@@ -4,7 +4,7 @@
 (defn flatten-step-result-outputs
   "Takes a nested step-result map (like those returned by a step or sent in a:pipeline-finished event)
   and converts it into a flat map of step results where every step is accessible with its step-id:
-  ```
+  ```clojure
   > (flatten-step-result-outputs {[1] {:status :success}
                                   [2] {:status  :success
                                        :outputs {[1 2] {:status :success :step [1 2]}}}})